+++ src/main/java/net/sodiumstudio/befriendmobs/entity/befriending/handlerpreset/HandlerItemGivingCount.java
@@ -128,10 +128,10 @@ public abstract class HandlerItemGivingCount extends HandlerItemGiving
 	public void serverTick(Mob mob)
 	{
 		mob.getCapability(BefMobCapabilities.CAP_BEFRIENDABLE_MOB).ifPresent((l) -> {
-			for (String key: l.getNbt().getAllKeys())
+			/*for (String key: l.getNbt().getAllKeys())
 			{
 
-			}
+			}*/
 			
 			if (!shouldIgnoreHatred())
 			{